Improved default module export
