+++ library/engine/mod.ts
@@ -1,2 +1,2 @@
-export { QueryEngine } from "./query_engine.ts";
-export { QueryEngineProxy } from "./query_engine_proxy.ts";
+export * from "./query_engine.ts";
+export * from "./types.ts";
